IO root PE list without MPI
This patch replaces the MPI-based method of generating the list of IO
domain root PEs with a method which locally tests each IO domain for
NULL_PEs and rebuilds the list locally inside of each rank's domain
object.

This eliminates the need for an `mpp_alltoall` call, which can be
expensive on some platforms.

This patch also adds some methods to query the domain object for this IO
root PE list.

We also added a quicksort method to re-sort the list of PEs, since most
pelist-based arguments require a sorted PE list.  This was added at the top
level of FMS as a utility.